fix:deleteMessageModalを削除してhoverMenuに直接書くのに変更(#70)
