+++ src/routes/channel/[id]/hoverMenu.svelte
@@ -1,14 +1,48 @@
 <script lang="ts">
   import { IconDots } from "@tabler/icons-svelte";
-  import DeleteMessageModal from "./deleteMessageModal.svelte";
-  import { onMount } from "svelte";
-
+  import {toastStore} from "$lib/store/toast";
   export let messageId: string = "";
   export let hoverMessageId: string = "";
   export let isLast: boolean = false;
+  import { repositoryFactory } from "$lib/repositories/RepositoryFactory";
 
+  const messageRepository = repositoryFactory.get("message");
   let isDisplayDeleteModal: boolean = false;
   let isHoverDropDown: boolean = false;
+  let deleteModal:HTMLDialogElement;
+
+  const deleteMessage = async (messageId :string)=>{
+    console.log("deleting");
+    await messageRepository
+      .deleteMessage(messageId)
+      .then((response) => {
+        console.log("メッセージ削除成功",response);
+        deleteModal.close();
+        toastStore.update((toast) =>{
+          return [
+            ...toast,
+            {
+              message: "メッセージを削除しました。",
+              type: "success",
+            }
+          ]
+        })
+      })
+      .catch((response) =>{
+        console.log("メッセージ削除失敗",response);
+        deleteModal.close();
+        toastStore.update((toast) =>{
+          return [
+            ...toast,
+            {
+              message: "メッセージを削除できませんでした。",
+              type: "error",
+            },
+          ];
+        });
+      });
+  }
+
 </script>
 
 <div class="relative">
@@ -43,7 +77,11 @@
         <li>
           <button
             class="text-error"
-            on:click={() => (isDisplayDeleteModal = true)}
+            on:click={() =>{
+               console.log("OnClick Delete id:"+ hoverMessageId +"isDisplayDeleteModal:"+ isDisplayDeleteModal);
+               isDisplayDeleteModal = true;
+               deleteModal.showModal();
+            }}
             >メッセージを削除</button
           >
         </li>
@@ -53,5 +91,17 @@
 </div>
 
 {#if isDisplayDeleteModal}
-  <DeleteMessageModal {messageId} />
+  <dialog bind:this={deleteModal} id="deleteModal" class="modal modal-bottom sm:modal-middle">
+    <div class="modal-box">
+      <h3 class="text-lg font-bold">メッセージを削除</h3>
+      <p class="py-4">メッセージを削除します。よろしいですか?</p>
+      <div class="modal-action">
+        <form method="dialog">
+          <!-- if there is a button in form, it will close the modal -->
+          <button class="btn">キャンセル</button>
+        </form>
+        <button class="btn" on:click={()=> deleteMessage(messageId)}>削除する</button>
+      </div>
+    </div>
+  </dialog>
 {/if}
